+++ wlanpi_webui/__version__.py
@@ -12,6 +12,6 @@ __description__ = "a custom WebUI made for the WLAN Pi"
 __url__ = "https://github.com/joshschmelzle/wlanpi-webui"
 __author__ = "Josh Schmelzle"
 __author_email__ = "josh@joshschmelzle.com"
-__version__ = "1.0.0-beta1"
+__version__ = "1.0.0-beta2"
 __status__ = "alpha"
 __license__ = "BSD-3-Clause"
