--- wlanpi_webui/fpms/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint("fpms", __name__)
-
-from wlanpi_webui.fpms import networkinfo
