+++ wlanpi_webui/network/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("network", __name__)
+
+from wlanpi_webui.network import network
