+++ wlanpi_webui/templates/public/layout.html
@@ -49,10 +49,10 @@
             <div class="uk-navbar-left">
 <a class="uk-navbar-item uk-logo" herf="#"><img src="{{ url_for('static', filename='img/favicon-96x96.png') }}" alt="WLAN Pi logo" style="width:64px;"></a>
               <ul class="uk-navbar-nav uk-visible@m">
-                <li><a href="{{ url_for('speedtest.speedtest') }}">Speedtest</a></li>
-                <li><a href="{{ url_for('profiler.profiler') }}">Profiler</a></li>
-                <li><a href="{{ url_for('fpms.networkinfo') }}">FPMS</a></li>
-                <li><a href="{{ url_for('cockpit') }}">Admin Panel (Cockpit)</a></li>
+                <li><a href="{{ url_for('speedtest.speedtest') }}">SPEEDTEST</a></li>
+                <li><a href="{{ url_for('profiler.profiler') }}">PROFILER</a></li>
+                <li><a href="{{ url_for('network.network') }}">NETWORK</a></li>
+                <li><a href="{{ url_for('admin') }}">ADMIN</a></li>
               </ul>
             </div>
               <div class="uk-navbar-right"> 
@@ -102,10 +102,10 @@
           <div class="uk-offcanvas-bar uk-offcanvas-bar-animation uk-offcanvas-slide">
             <button class="uk-offcanvas-close uk-close uk-icon" type="button" data-uk-close></button>
             <ul class="uk-nav uk-nav-default">
-              <li><a href="{{ url_for('speedtest.speedtest') }}">Speedtest</a></li>
-              <li><a href="{{ url_for('profiler.profiler') }}">Profiler</a></li>
-              <li><a href="{{ url_for('fpms.networkinfo') }}">FPMS</a></li>
-              <li><a href="{{ url_for('cockpit') }}">Admin Panel (Cockpit)</a></li> 
+             <li><a href="{{ url_for('speedtest.speedtest') }}">SPEEDTEST</a></li>
+              <li><a href="{{ url_for('profiler.profiler') }}">PROFILER</a></li>
+              <li><a href="{{ url_for('network.network') }}">NETWORK</a></li>             
+              <li><a href="{{ url_for('admin') }}">ADMIN</a></li> 
             </ul>
           </div>
         </div>
