+++ wlanpi_webui/templates/public/network.html
@@ -1,6 +1,6 @@
 {% extends "public/layout.html" %} {% block content %}
 <div class="uk-flex uk-padding uk-flex-column">
-<div><b>FPMS</b> network information scripts</p></div>
+<!--<div><p>a subset of network information scripts</p></div>-->
 <div class="uk-grid uk-child-width-1-3@m" uk-grid uk-height-match="target: > div > .uk-card">
     <div>   
       <div class="uk-card uk-card-default uk-card-medium uk-card-hover uk-card-body">
